Improved status messages

+++ fabfile.py
@@ -1,10 +1,11 @@
-from fabric.api import run, sudo, cd, settings, task, put, prompt
+from fabric.api import run, sudo, cd, settings, task, put, prompt, abort
 from fabric.colors import yellow, green
 from fabric.contrib.files import exists
+from fabric.contrib.console import confirm
 from StringIO import StringIO
 from configs import nginx_server_config, nginx_client_config, nginx_ssl_config, openseedbox_server_config, openseedbox_client_config
 
-packages = "unzip git openjdk-6-jre-headless python-software-properties build-essential libssl-dev libpcre3-dev nginx"
+packages = "unzip git openjdk-6-jre-headless python-software-properties build-essential libssl-dev libpcre3-dev"
 server_only_packages = "transmission-daemon cryptsetup"
 play_location = "http://download.playframework.org/releases/play-1.2.5.zip"
 nginx_location = "http://nginx.org/download/nginx-1.2.6.tar.gz"
@@ -18,11 +19,51 @@ openseedbox_backend_port=9001
 openseedbox_client_port=9000
 openseedbox_backend_path="/media/openseedbox"
 openseedbox_ssl_enabled=True
-openseedbox_client_db_url="jdbc:mysql://root:S*a!m()$@localhost/openseedbox"
+openseedbox_client_db_url="jdbc:mysql:sites:sitespassword@localhost/openseedbox"
 
 @task
-def bootstrap(type="server", server_name="localhost", server_api_key=""):
+def bootstrap(type="server", server_name="localhost", server_api_key="", configure_nginx=True, encrypt=True):
+	"""
+	Bootstraps a new OpenSeedbox node
+	
+	Args:
+		type:
+		either "client" or "server". "client" is for the frontend
+		website node, "server" is for the backend torrenting nodes.
+		
+		server_name:
+		defaults to "localhost". This is what is put in the webserver
+		config file.
+		
+		server_api_key:
+		Only applies when type is set to "server". This is written to the Play!
+		config file and is used by the client to communicate with this server.
+		
+		configure_nginx: 
+		defaults to True. When set, a custom nginx will be downloaded and
+		compiled, and nginx config files will be written out.
+		
+		encrypt:
+		Only applies when type is set to "server". Defaults to True. When True,
+		a dm-crypt encrypted mountpoint will be created to store torrent data
+		on.
+	"""
+	#Fabric passes everything as strings, so make sure we can use the following properly in IF statements
+	configure_nginx = str2bool(configure_nginx)
+	encrypt = str2bool(encrypt)
+	folder = get_folder(type)
+	
 	print(green("Bootstrapping %s node" % type))
+	text("""
+Node Type: %s
+Server Name: %s
+Server API Key: %s
+Configure NGINX: %s
+Configure Encrypted Folder: %s""" % (type, server_name, server_api_key, configure_nginx, encrypt))
+			 
+	if not confirm("Are these settings all good?"):
+		abort("Aborting to prevent injury.")
+	
 	step = 1
 	text("%s. Creating /src directory" % step)
 	sudo("mkdir -p /src")
@@ -33,16 +74,19 @@ def bootstrap(type="server", server_name="localhost", server_api_key=""):
 	
 	text("%s. Updating apt package repo and installing required packages" % step)
 	sudo("apt-get update -qq")
-	all_packages = "%s %s" % (packages, server_only_packages) if (type=="server") else packages
+	all_packages = "%s %s" % (packages, server_only_packages) if type == "server" else packages
+	if configure_nginx:
+		all_packages += " nginx"
 	sudo("apt-get install -qqy %s" % all_packages)
-	if (type == "server"):
+	if type == "server":
 		#We cant install updated transmission from ppa until python-software-properties was installed in the previous step
 		sudo("apt-add-repository -y ppa:transmissionbt/ppa")
 		sudo("apt-get update -qq")
 		sudo("apt-get install -qqy transmission-daemon")
-	with settings(warn_only=True):
-		sudo("apt-mark hold nginx") #incase there is an nginx update in the repos, we still want to use the version we compile below
-	step += 1
+	if configure_nginx:
+		with settings(warn_only=True):
+			sudo("apt-mark hold nginx") #incase there is an nginx update in the repos, we still want to use the version we compile below
+		step += 1
 		
 	with cd("/src"):
 		text("%s. Installing Play 1.2.5" % step)
@@ -59,10 +103,10 @@ def bootstrap(type="server", server_name="localhost", server_api_key=""):
 		step += 1
 		
 		text("%s. Checking out git repositories" % step)
-		pull_or_clone()
+		pull_or_clone(type)
 		step += 1
 	
-	if (type == "server"):	
+	if type == "server" and encrypt:	
 		if not exists("~/data.img"):
 			text("%s. Creating encrypted partion" % step)
 			size = prompt("How many GB do you want the partition to be?")
@@ -80,35 +124,36 @@ def bootstrap(type="server", server_name="localhost", server_api_key=""):
 		sudo("chown -R %s /media/openseedbox" % user)
 		step += 1
 	
+	if type == "server":
 		text("%s. Removing transmission-daemon from startup" % step)
 		sudo("update-rc.d -f transmission-daemon remove")
 		step += 1
-		
-	text("%s. Downloading and compiling custom nginx" % step) #note: we installed nginx via apt-get before because we wanted to get the init.d scripts
-	sudo("service nginx stop", pty=False)
-	with cd("/src"):
-		run("wget -q %s" % nginx_location)
-		run("tar -xvf nginx*.tar.gz")
-		run("rm nginx*.tar.gz")
-		if not exists("headers-more-nginx-module"):
-			run("git clone -q https://github.com/agentzh/headers-more-nginx-module.git") #get headers-more nginx module
-		else:
-			run("cd headers-more-nginx-module && git pull")
-		if not exists("mod_zip"):
-			run("git clone -q https://github.com/evanmiller/mod_zip.git") #get mod_zip ngnx module
-		else:
-			run("cd mod_zip && git pull")			
-	with cd("/src/nginx*"):
-		run("./configure --with-http_ssl_module --add-module=/src/mod_zip/ --prefix=/etc/nginx --conf-path=/etc/nginx/nginx.conf --error-log-path=/var/log/nginx/error.log --pid-path=/var/run/nginx.pid --http-log-path=/var/log/nginx/access.log --lock-path=/var/lock/nginx.lock --sbin-path=/usr/sbin/nginx --add-module=/src/headers-more-nginx-module")
-		run("make")
-		sudo("make install")
-	sudo("service nginx start", pty=False)
-	step += 1
 	
-	folder = get_folder(type)
-	text("%s. Creating nginx config for %s" % (step, folder))
-	create_nginx_config(type, server_name)
-	step += 1
+	if configure_nginx:
+		text("%s. Downloading and compiling custom nginx" % step) #note: we installed nginx via apt-get before because we wanted to get the init.d scripts
+		sudo("service nginx stop", pty=False)
+		with cd("/src"):
+			run("wget -q %s" % nginx_location)
+			run("tar -xvf nginx*.tar.gz")
+			run("rm nginx*.tar.gz")
+			if not exists("headers-more-nginx-module"):
+				run("git clone -q https://github.com/agentzh/headers-more-nginx-module.git") #get headers-more nginx module
+			else:
+				run("cd headers-more-nginx-module && git pull")
+			if not exists("mod_zip"):
+				run("git clone -q https://github.com/evanmiller/mod_zip.git") #get mod_zip ngnx module
+			else:
+				run("cd mod_zip && git pull")			
+		with cd("/src/nginx*"):
+			run("./configure --with-http_ssl_module --add-module=/src/mod_zip/ --prefix=/etc/nginx --conf-path=/etc/nginx/nginx.conf --error-log-path=/var/log/nginx/error.log --pid-path=/var/run/nginx.pid --http-log-path=/var/log/nginx/access.log --lock-path=/var/lock/nginx.lock --sbin-path=/usr/sbin/nginx --add-module=/src/headers-more-nginx-module")
+			run("make")
+			sudo("make install")
+		sudo("service nginx start", pty=False)
+		step += 1
+			
+		text("%s. Creating nginx config for %s" % (step, folder))
+		create_nginx_config(type, server_name)
+		step += 1
 	
 	text("%s. Creating play config for %s" % (step, folder))
 	create_play_config(type, server_api_key)
@@ -118,15 +163,39 @@ def bootstrap(type="server", server_name="localhost", server_api_key=""):
 	start_play(type)
 
 @task
-def update_code():
-	print(green("Updating code"))
+def update_code(type="server"):
+	"""
+	Updates the OpenSeedbox code and restarts the Play! services
+	
+	Args:
+		type:
+		Either "client" or "server" (defaults to "server"). This sets what
+		git repositories to update.
+	"""
+	print(green("Updating code for type: %s" % type))
 	with cd("/src"):
 		pull_or_clone()
+	print(green("Restarting play for type: %s" % type))
+	start_play(type) #start_play calls stop_play first
 		
 @task
-def start_play(type="server"):
+def start_play(type="server", encrypt=True):
+	"""
+	Stops and then starts the Play! services
+	
+	Args:
+		type:
+		Either "client" or "server" (defaults to "server"). This sets what
+		Play! services to restart.
+		
+		encrypt:
+		Only applies when type == "server". Defaults to True. If True, the
+		encrypted partition will be remounted if it is not mounted.
+	"""
 	folder = get_folder(type)
 	stop_play(type)
+	if type == "server" and encrypt:
+		mount_encrypted_partition()
 			
 	with cd("/src/%s" % folder):
 		run("play start --%prod", pty=False)
@@ -134,7 +203,7 @@ def start_play(type="server"):
 	text("%s started!" % folder)
 	
 @task
-def update_servers():
+def update_servers(type="server"):
 	text("Updating apt package cache and packages")
 	sudo("apt-get update -qq")
 	sudo("apt-get upgrade -qqy")
@@ -143,7 +212,8 @@ def update_servers():
 	with cd("/src"):
 		pull_or_clone()		
 		
-	text("Restarting nginx and play services")
+	text("Restarting play services")
+	start_play(type)
 	
 @task
 def update_configs(type="server", server_name="localhost", server_api_key=""):
@@ -169,16 +239,30 @@ def stop_play(type="server"):
 				run("kill -9 %s" % pid)	
 				run("rm server.pid")	
 	
-def pull_or_clone():
+def pull_or_clone(type="server"):
 	#note: assumes caller is using 'with cd("/src"):'
 	for r in repositories:
 		name = r[0]
 		repo = r[1]
+		
+		#skip getting unnecessary repos
+		if type == "server" and name == "openseedbox":
+			continue
+		if type == "client" and name == "openseedbox-server":
+			continue
+			
 		if not exists(name):
 			run("git clone -q %s /src/%s" % (repo, name))
 		else:
 			with cd("/src/%s" % name):
+				#move application.conf elsewhere because otherwise there will be git pull conflicts
+				moved_application_conf=False
+				if exists("conf/application.conf"):
+					run("mv conf/application.conf /tmp/%s.application.conf" % name)
+					moved_application_conf=True
 				run("git pull")
+				if moved_application_conf:
+					run("mv /tmp/%s.application.conf conf/application.conf" % name)
 		with cd("/src/%s" % name):
 			run("play deps --sync")
 				
@@ -219,3 +303,9 @@ def get_config_params(server_name=""):
 	
 def text(t):
 	print(yellow(t))
+	
+def str2bool(v):
+	if isinstance(v, bool):
+		return v;
+	return v.lower() in ("yes", "true", "t", "1")
+
